+++ packages/editor/editor-core/examples/5-full-page.tsx
@@ -11,6 +11,7 @@ import {
   storyMediaProviderFactory,
   storyContextIdentifierProviderFactory,
   macroProvider,
+  cardProvider,
 } from '@atlaskit/editor-test-helpers';
 import { mention, emoji, taskDecision } from '@atlaskit/util-data-test';
 import { MockActivityResource } from '@atlaskit/activity/dist/es5/support';
@@ -147,7 +148,9 @@ export class ExampleEditor extends React.Component<Props, State> {
             UNSAFE_allowLayouts={true}
             allowGapCursor={true}
             allowTemplatePlaceholders={{ allowInserting: true }}
-            UNSAFE_allowCards={true}
+            UNSAFE_cards={{
+              provider: Promise.resolve(cardProvider),
+            }}
             {...providers}
             media={{ provider: mediaProvider, allowMediaSingle: true }}
             placeholder="Write something..."