+++ packages/media/media-viewer/__tests__/newgen/viewers/doc/index.spec.tsx
@@ -7,7 +7,7 @@ import { Subject } from 'rxjs/Subject';
 import { FileItem } from '@atlaskit/media-core';
 import { Stubs } from '../../../_stubs';
 import { Spinner } from '../../../../src/newgen/loading';
-import { PDFViewer } from '../../../../src/newgen/viewers/pdf/index';
+import { DocViewer } from '../../../../src/newgen/viewers/doc/index';
 
 function createContext() {
   const subject = new Subject<FileItem>();
@@ -30,18 +30,18 @@ function createFixture(fetchPromise, item, collectionName?) {
   const context = createContext();
   const onClose = jest.fn(() => fetchPromise);
   const el = mount(
-    <PDFViewer item={item} context={context} collectionName={collectionName} />,
+    <DocViewer item={item} context={context} collectionName={collectionName} />,
   );
   el.instance()['fetch'] = jest.fn();
   return { context, el, onClose };
 }
 
-describe('PDFViewer', () => {
+describe('DocViewer', () => {
   afterEach(() => {
     constructAuthTokenUrlSpy.mockClear();
   });
 
-  it('assigns a document object when successful', async () => {
+  it('assigns a document src when successful', async () => {
     const fetchPromise = Promise.resolve();
     const item: FileItem = {
       type: 'file',
@@ -60,7 +60,7 @@ describe('PDFViewer', () => {
     await el.instance()['init']();
 
     expect(el.state()).toMatchObject({
-      doc: {
+      src: {
         status: 'SUCCESSFUL',
       },
     });
@@ -103,7 +103,7 @@ describe('PDFViewer', () => {
     await el.instance()['init']();
 
     expect(el.state()).toMatchObject({
-      doc: {
+      src: {
         status: 'FAILED',
         err: new Error('no pdf artifacts found for this file'),
       },