+++ packages/media/media-viewer/src/newgen/viewers/doc/pdfRenderer.tsx
@@ -5,6 +5,7 @@ import { injectGlobal } from 'styled-components';
 import { ZoomControls } from '../../zoomControls';
 import { PDFWrapper } from '../../styled';
 import { closeOnDirectClick } from '../../utils/closeOnDirectClick';
+import { Outcome } from '../../domain';
 
 export const pdfViewerClassName = 'pdfViewer';
 
@@ -36,30 +37,41 @@ injectGlobal`
 
 pdfjsLib.GlobalWorkerOptions.workerSrc = '/'; // TODO: use web workers instead of fake worker.
 
-export const fetch = async (url: string): Promise<Blob> => {
+const fetch = async (url: string): Promise<Blob> => {
   return pdfjsLib.getDocument(url).promise;
 };
 
 export type Props = {
-  doc: any;
+  src: string;
   onClose?: () => void;
 };
 
 export type State = {
+  doc: Outcome<any, Error>;
   zoom: number;
 };
 
-const initialState: State = { zoom: 100 };
+const initialState: State = { zoom: 100, doc: { status: 'PENDING' } };
 
-export class PDFViewer extends React.PureComponent<Props, State> {
+export class PDFRenderer extends React.PureComponent<Props, State> {
   private el: HTMLDivElement;
   private pdfViewer: any;
 
   state: State = initialState;
 
   componentDidMount() {
+    this.init();
+  }
+
+  private async init() {
     this.pdfViewer = new PDFJSViewer.PDFViewer({ container: this.el });
-    this.pdfViewer.setDocument(this.props.doc);
+    try {
+      const doc = await fetch(this.props.src);
+      this.setState({ doc: { status: 'SUCCESSFUL', data: doc } });
+      this.pdfViewer.setDocument(doc);
+    } catch (err) {
+      this.setState({ doc: { status: 'FAILED', err } });
+    }
   }
 
   private savePdfElement = el => {