+++ packages/search/global-search/src/components/home/HomeQuickSearchContainer.tsx
@@ -201,6 +201,7 @@ export class HomeQuickSearchContainer extends React.Component<Props, State> {
       jiraResults,
       confluenceResults,
       peopleResults,
+      searchSessionId,
     } = this.state;
 
     return (
@@ -210,6 +211,7 @@ export class HomeQuickSearchContainer extends React.Component<Props, State> {
         isLoading={isLoading}
         query={query}
         linkComponent={linkComponent}
+        searchSessionId={searchSessionId}
       >
         {renderSearchResults({
           query,